ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.4.3
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
psycopg2-binary==2.9.9
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
